ref: move interfaces to base modules
