+++ contracts/src/finance/vesting_wallet.rs
@@ -30,7 +30,7 @@ use stylus_sdk::{
     call::{self, call, Call},
     contract, evm, function_selector,
     storage::TopLevelStorage,
-    stylus_proc::{public, sol_storage, SolidityError},
+    stylus_proc::{public, sol_interface, sol_storage, SolidityError},
 };
 
 use crate::{
@@ -79,14 +79,11 @@ pub enum Error {
     InvalidToken(InvalidToken),
 }
 
-pub use token::IErc20;
-#[allow(missing_docs)]
-mod token {
-    stylus_sdk::stylus_proc::sol_interface! {
-        /// Interface of the ERC-20 token.
-        interface IErc20 {
-            function balanceOf(address account) external view returns (uint256);
-        }
+sol_interface! {
+    /// Interface of the ERC-20 token.
+    #[allow(missing_docs)]
+    interface IErc20 {
+        function balanceOf(address account) external view returns (uint256);
     }
 }
