+++ contracts/src/token/erc721/mod.rs
@@ -9,6 +9,7 @@ use stylus_sdk::{
     call::{self, Call, MethodError},
     evm, msg,
     prelude::*,
+    stylus_proc::sol_interface,
 };
 
 use crate::utils::{
@@ -160,29 +161,26 @@ impl MethodError for Error {
     }
 }
 
-pub use receiver::IERC721Receiver;
-#[allow(missing_docs)]
-mod receiver {
-    stylus_sdk::stylus_proc::sol_interface! {
-        /// [`Erc721`] token receiver interface.
+sol_interface! {
+    /// [`Erc721`] token receiver interface.
+    ///
+    /// Interface for any contract that wants to support `safe_transfers`
+    /// from [`Erc721`] asset contracts.
+    #[allow(missing_docs)]
+    interface IERC721Receiver {
+        /// Whenever an [`Erc721`] `token_id` token is transferred
+        /// to this contract via [`Erc721::safe_transfer_from`].
         ///
-        /// Interface for any contract that wants to support `safe_transfers`
-        /// from [`Erc721`] asset contracts.
-        interface IERC721Receiver {
-            /// Whenever an [`Erc721`] `token_id` token is transferred
-            /// to this contract via [`Erc721::safe_transfer_from`].
-            ///
-            /// It must return its function selector to confirm the token transfer.
-            /// If any other value is returned or the interface is not implemented
-            /// by the recipient, the transfer will be reverted.
-            #[allow(missing_docs)]
-            function onERC721Received(
-                address operator,
-                address from,
-                uint256 token_id,
-                bytes calldata data
-            ) external returns (bytes4);
-        }
+        /// It must return its function selector to confirm the token transfer.
+        /// If any other value is returned or the interface is not implemented
+        /// by the recipient, the transfer will be reverted.
+        #[allow(missing_docs)]
+        function onERC721Received(
+            address operator,
+            address from,
+            uint256 token_id,
+            bytes calldata data
+        ) external returns (bytes4);
     }
 }
